Partial revert of the socket API changes to realign with master
This is closer to master socket API code and doesn't break compatibility
with older shell integration plugins. To this effect, it removes all
the Windows specific API in favor of keeping only the two extra commands
for offline/online state which were apparently used on the Mac side
only.

Also this simplifies quite a bit the use of mapping since most paths
were going through FileData::get anyway which does the proper mapping.

Signed-off-by: Kevin Ottens <kevin.ottens@nextcloud.com>